pyproject.toml
[project]
name = "classificador"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "ipykernel>=7.0.1",
    "matplotlib>=3.10.7",
    "nltk>=3.9.2",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "xgboost>=3.1.1",
]
